Add photo draft

+++ src/main/java/com/revature/repository/AlbumRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.revature.model.Album;
 
-public interface AlbumRepository extends JpaRepository<Album, Integer> {
+public interface AlbumRepository extends JpaRepository<Album, Long> {
 
 	Album findAlbumByAlbumName(String albumName);
